Convert managed purl script to stimulus
Fixes #5364
