+++ app/javascript/application.js
@@ -27,7 +27,6 @@ import "./eds_range_limit";
 import "./exhibitPanel";
 import "./facet-options-checkboxes";
 import "./feedback_form";
-import "./jquery.managed-purl";
 import "./location-hours";
 import "./purl-embed";
 import "./recent-selections";
